get student | teacher | admin | uni | college | depertment page for both admin and client side
