+++ server/routes/admin.routes.js
@@ -2,21 +2,35 @@ import { Router } from "express";
 // Middlewares
 import authinticate from "../middlewares/authintication.middleware.js";
 import authorize from "../middlewares/authorization.middleware.js";
-import {
-  createStudent,
-  createAdmin,
-} from "../controllers/admin.controller.js";
+import { createAdmin } from "../controllers/admin.controller.js";
 import {
   getAdminById,
-  createTeacher,
-  updateTeacher,
   updateAdmin,
   deleteAdmin,
   getPageOfAdmins,
-  createCollege,
 } from "../controllers/admin.controller.js";
-import { createUniversity, updateUniversity } from "../controllers/university.controller.js";
-import { updateCollege } from "../controllers/college.controller.js";
+import {
+  createStudent,
+  deleteStudent,
+  updateStudent,
+} from "../controllers/student.controller.js";
+import {
+  createTeacher,
+  deleteTeacher,
+  updateTeacher,
+} from "../controllers/teacher.controller.js";
+import {
+  createUniversity,
+  updateUniversity,
+} from "../controllers/university.controller.js";
+import {
+  updateCollege,
+  createCollege,
+} from "../controllers/college.controller.js";
+import {
+  validateObjectId,
+  validateStudentUpdateData,
+} from "../middlewares/validation.middleware.js";
 
 const router = Router();
 
@@ -31,16 +45,11 @@ router
   .delete(authinticate, authorize(["super-admin"]), deleteAdmin);
 
 router
-  .route("/:id")
-  .get(authinticate, authorize(["super-admin"]), getAdminById);
-
-router
   .route("/student")
   .post(authinticate, authorize(["admin", "super-admin"]), createStudent);
 router
   .route("/teacher")
-  .post(authinticate, authorize(["admin", "super-admin"]), createTeacher)
-  .patch(authinticate, authorize(["admin", "super-admin"]), updateTeacher);
+  .post(authinticate, authorize(["admin", "super-admin"]), createTeacher);
 
 router
   .route("/university")
@@ -51,5 +60,23 @@ router
   .route("/university/college")
   .post(authinticate, authorize(["admin", "super-admin"]), createCollege)
   .patch(authinticate, authorize(["admin", "super-admin"]), updateCollege);
+
+router
+  .route("/:id")
+  .get(authinticate, authorize(["super-admin"]), getAdminById);
+
+router
+  .route("/teacher/:id")
+  .patch(authinticate, authorize(["admin", "super-admin"]), updateTeacher)
+  .delete(authinticate, authorize(["admin", "super-admin"]), deleteTeacher);
+
+router
+  .route("/student/:id")
+  .patch(
+    authinticate,
+    authorize(["admin", "super-admin"]),
+    updateStudent
+  )
+  .delete(authinticate, authorize(["admin", "super-admin"]), deleteStudent);
 
 export default router;
